UserToDto 로직 UserService 생성 및 회원정지 메서드 구현 중 user id 형태 long으로 변경

+++ src/main/java/org/sbsplus/util/DataCreator.java
@@ -206,7 +206,7 @@ public class DataCreator {
     }
 
     protected User randomUser(){
-        return userRepository.findById(random.nextInt(userNum) + 1).orElse(null);
+        return userRepository.findById(random.nextLong(userNum) + 1).orElse(null);
     }
 
     protected Article randomArticle(){
